+++ index.js
@@ -4,17 +4,17 @@ export { CalloutContext, useCallout } from './src/CalloutContext';
 
 /* internal utilities */
 export { stripesShape } from './src/Stripes';
-export { withStripes, useStripes } from './src/StripesContext';
+export { withStripes, useStripes, StripesContext } from './src/StripesContext';
 export { useModules } from './src/ModulesContext';
 export { withModule, withModules } from './src/components/Modules';
 export { default as stripesConnect } from './src/stripesConnect';
 export { default as Pluggable } from './src/Pluggable';
-export { setServicePoints, setCurServicePoint } from './src/loginServices';
-export { updateUser } from './src/loginServices';
+export { updateUser, updateTenant, validateUser } from './src/loginServices';
 export { default as coreEvents } from './src/events';
 export { default as useOkapiKy } from './src/useOkapiKy';
 export { default as withOkapiKy } from './src/withOkapiKy';
 export { default as useCustomFields } from './src/useCustomFields';
+export { default as createReactQueryClient } from './src/createReactQueryClient';
 
 /* components */
 export { default as AppContextMenu } from './src/components/MainNav/CurrentApp/AppContextMenu';
@@ -27,14 +27,25 @@ export { default as AppIcon } from './src/components/AppIcon';
 export { Route, Switch, Redirect } from './src/components/NestedRouter';
 export {
   ModuleHierarchyContext,
+  ModuleHierarchyProvider,
   useModuleHierarchy,
   useNamespace,
   withNamespace,
+  LastVisitedContext,
+  withLastVisited,
 } from './src/components';
+
+/* Queries */
+export { useChunkedCQLFetch } from './src/queries';
 
 /* misc */
 export { supportedLocales } from './src/loginServices';
 export { supportedNumberingSystems } from './src/loginServices';
 export { userLocaleConfig } from './src/loginServices';
+export * from './src/consortiaServices';
 export { default as queryLimit } from './src/queryLimit';
 export { default as init } from './src/init';
+
+export { registerServiceWorker, unregisterServiceWorker } from './src/serviceWorkerRegistration';
+
+export { getEventHandler } from './src/handlerService';