+++ src/okapiReducer.js
@@ -25,8 +25,10 @@ export default function okapiReducer(state = {}, action) {
     case 'CLEAR_CURRENT_USER':
       return Object.assign({}, state, { currentUser: {}, currentPerms: {} });
     case 'SET_SESSION_DATA': {
-      const { perms, user, token } = action.session;
-      return { ...state, currentUser: user, currentPerms: perms, token };
+      const { perms, user, token, tenant } = action.session;
+      const sessionTenant = tenant || state.tenant;
+
+      return { ...state, currentUser: user, currentPerms: perms, token, tenant: sessionTenant };
     }
     case 'SET_AUTH_FAILURE':
       return Object.assign({}, state, { authFailure: action.message });
@@ -38,10 +40,6 @@ export default function okapiReducer(state = {}, action) {
       return Object.assign({}, state, { okapiReady: true });
     case 'SERVER_DOWN':
       return Object.assign({}, state, { serverDown: true });
-    case 'SET_CURRENT_SERVICE_POINT':
-      return { ...state, currentUser: { ...state.currentUser, curServicePoint: action.servicePoint } };
-    case 'SET_USER_SERVICE_POINTS':
-      return { ...state, currentUser: { ...state.currentUser, servicePoints: action.servicePoints } };
     case 'UPDATE_CURRENT_USER':
       return { ...state, currentUser: { ...state.currentUser, ...action.data } };
     default: