+++ test/bigtest/helpers/setup-application.js
@@ -1,14 +1,14 @@
-import ReactDOM from 'react-dom';
 import { beforeEach } from '@bigtest/mocha';
-import { setupAppForTesting, visit, location } from '@bigtest/react';
 import localforage from 'localforage';
+
 import { reset } from '@folio/stripes-connect';
+import { visit, location } from '@folio/stripes-testing/bigtest';
 
 // load these styles for our tests
 import '@folio/stripes-components/lib/global.css';
 
+import { setupAppForTesting } from './setupAppForTesting';
 import startMirage from '../network/start';
-
 import App from '../../../src/App';
 
 import {
@@ -96,9 +96,6 @@ export default function setupApplication({
         this.app = null;
       }
     });
-
-    // set the root to 100% height
-    document.getElementById(mountId).style.height = '100%';
 
     // setup react validators
     Object.defineProperties(this, {